+++ dotnet/test/SmartStandards.Logging.Tests/Logging/SmartStandards/Transport/TraceBusFeed.Buffer.Tests.cs
@@ -1,0 +1,33 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
+using System.Threading;
+
+namespace Logging.SmartStandards.Transport {
+
+  //[TestClass]
+  //public class TraceBusFeedBufferTests {
+
+  //  [TestMethod()]
+  //  public void MyCircularBuffer_AutoFlush_ShouldInvokeCallbackFrequently() {
+
+  //    var buffer = new TraceBusFeed.MyCircularBuffer<TraceBusFeed.QueuedEvent>(100);
+
+  //    buffer.StartAutoFlush(() => { Console.WriteLine("■AUTOFLUSH 1■"); }, 1000);
+
+  //    Thread.Sleep(1100);
+
+  //    buffer.StartAutoFlush(() => { Console.WriteLine("■AUTOFLUSH 1 - before stopping■"); }, 1000);
+
+  //    Thread.Sleep(1100);
+
+  //    buffer.StopAutoFlush();
+
+  //    Thread.Sleep(1100);
+
+  //    buffer.StartAutoFlush(() => { Console.WriteLine("■AUTOFLUSH 2■"); }, 1000);
+
+  //    Thread.Sleep(1100);
+  //  }
+
+  //}
+}